requirements.txt
azure-common==1.1.26
azure-storage-common==2.1.0
azure-storage-queue==2.1.0
Flask==1.1.2
oauthlib==3.1.0
requests==2.22.0
requests-oauthlib==1.3.0
urllib3==1.25.3
